// packages/build/src/types.ts
export type SWTargets = 'baseline-widely-available' | 'esnext' | string | string[]
// ...
export type SWTarget = SWTargets | {
  /**
   * Classic service worker build target.
   */
  classic: SWTargets
  /**
   * Module service worker build target.
   */
  module: SWTargets
}
// ...
export type SWType = 'classic' | 'module' | 'classic-and-module'
// ...
export interface GlobPartial {
  /**
   * Determines whether or not symlinks are followed when generating the
   * precache manifest. For more information, see the definition of `followSymbolicLinks`
   * in the `glob` [documentation](https://superchupu.dev/tinyglobby).
   * @default true
   */
  globFollow?: boolean
  /**
   * A set of patterns matching files to always exclude when generating the
   * precache manifest. For more information, see the definition of `ignore` in
   * the `glob` [documentation](https://superchupu.dev/tinyglobby).
   */
  globIgnores?: Array<string>
  /**
   * Files matching any of these patterns will be included in the precache
   * manifest. For more information, see the
   * [tinyglobby](https://superchupu.dev/tinyglobby).
   */
  globPatterns?: Array<string>
  /**
   * If true, an error reading a directory when generating a precache manifest
   * will cause the build to fail. If false, the problematic directory will be
   * skipped. For more information, see the definition of `strict` in the `glob`
   * [documentation](https://superchupu.dev/tinyglobby).
   * @default true
   */
  globStrict?: boolean
  /**
   * If a URL is rendered based on some server-side logic, its contents may
   * depend on multiple files or on some other unique string value. The keys in
   * this object are server-rendered URLs. If the values are an array of
   * strings, they will be interpreted as `glob` patterns, and the contents of
   * any files matching the patterns will be used to uniquely version the URL.
   * If used with a single string, it will be interpreted as unique versioning
   * information that you've generated for a given URL.
   */
  templatedURLs?: {
    [key: string]: string | Array<string>
  }
}
// ...
export interface RequiredSWDestPartial {
  /**
   * The path and filename of the service worker file that will be created by
   * the build process, relative to the current working directory. It must end
   * in '.js'.
   */
  swDest: string
}